refactor(ui): move 11 pure helpers from feed.ts to feed/helpers.ts
Second sub-commit of the feed/ split (codemem-ug38). Moves every
internally-used pure helper function out of feed.ts into
feed/helpers.ts, verbatim. Each function was copied byte-for-byte from
its original location, then prefixed with `export` so feed.ts can import
it. No bodies were rewritten, no logic changed.

Helpers moved:
 - mergeMetadata, extractFactsFromBody, sentenceFacts
 - isLowSignalObservation, itemSignature, itemKey
 - feedScopeLabel, trustStateLabel, authorLabel
 - mergeFeedItems, mergeRefreshFeedItems

Exports kept in feed.ts for now because they're part of its public API
surface (syncInspectorQueryDraft, parseInspectorWorkingSet,
packTraceContextKey) — those stay until a later sub-commit re-homes
them either as helper exports or as their own inspector/ module.

Mechanical move only. Gate stays green (tsc / lint / 1417 tests).
feed.ts drops further from 1483 to 1399 lines.

Left for follow-up sub-commits on the same bead:
 - pagination.ts (module-local offsets + loadMoreFeedPage)
 - card.tsx (FeedItemCard + FeedItemMenu + FeedViewToggle)
 - skeleton.tsx (FeedSkeletonItem)
 - list.tsx (FeedList)
 - inspector.tsx (ContextInspectorPanel + TraceCandidateGroup)
 - view.tsx (FeedTabView)
