+++ packages/ui/src/tabs/feed.ts
@@ -24,77 +24,22 @@ import { openSyncConfirmDialog } from "./sync/sync-dialogs";
 
 export type { FeedItem, FeedItemMetadata } from "./feed/types";
 
+import {
+	authorLabel,
+	extractFactsFromBody,
+	feedScopeLabel,
+	isLowSignalObservation,
+	itemKey,
+	itemSignature,
+	mergeFeedItems,
+	mergeMetadata,
+	mergeRefreshFeedItems,
+	sentenceFacts,
+	trustStateLabel,
+} from "./feed/helpers";
 import type { FeedItem, FeedItemMetadata, FeedSummary, ItemViewMode } from "./feed/types";
 
-/* ── Helpers ─────────────────────────────────────────────── */
-
-function mergeMetadata(metadata: unknown): FeedItemMetadata {
-	if (!metadata || typeof metadata !== "object") return {};
-	const meta = metadata as FeedItemMetadata;
-	const importMeta = meta.import_metadata;
-	if (importMeta && typeof importMeta === "object") {
-		return { ...importMeta, ...meta };
-	}
-	return meta;
-}
-
-function extractFactsFromBody(text: unknown): string[] {
-	if (!text) return [];
-	const lines = String(text)
-		.split("\n")
-		.map((l) => l.trim())
-		.filter(Boolean);
-	const bullets = lines.filter((l) => /^[-*\u2022]\s+/.test(l) || /^\d+\./.test(l));
-	if (!bullets.length) return [];
-	return bullets.map((l) => l.replace(/^[-*\u2022]\s+/, "").replace(/^\d+\.\s+/, ""));
-}
-
-function sentenceFacts(text: string, limit = 6): string[] {
-	const raw = String(text || "").trim();
-	if (!raw) return [];
-	const collapsed = raw.replace(/\s+/g, " ").trim();
-	const parts = collapsed
-		.split(/(?<=[.!?])\s+/)
-		.map((p) => p.trim())
-		.filter(Boolean);
-	const facts: string[] = [];
-	for (const part of parts) {
-		if (part.length < 18) continue;
-		facts.push(part);
-		if (facts.length >= limit) break;
-	}
-	return facts;
-}
-
-function isLowSignalObservation(item: FeedItem): boolean {
-	const title = normalize(item.title);
-	const body = normalize(item.body_text);
-	if (!title && !body) return true;
-	const combined = body || title;
-	if (combined.length < 10) return true;
-	if (title && body && title === body && combined.length < 40) return true;
-	const lead = title.charAt(0);
-	if ((lead === "\u2514" || lead === "\u203a") && combined.length < 40) return true;
-	if (title.startsWith("list ") && combined.length < 20) return true;
-	if (combined === "ls" || combined === "list ls") return true;
-	return false;
-}
-
-function itemSignature(item: FeedItem): string {
-	return String(
-		item.id ??
-			item.memory_id ??
-			item.observation_id ??
-			item.session_id ??
-			item.created_at_utc ??
-			item.created_at ??
-			"",
-	);
-}
-
-function itemKey(item: FeedItem): string {
-	return `${String(item.kind || "").toLowerCase()}:${itemSignature(item)}`;
-}
+/* ── Constants + module state ─────────────────────────────── */
 
 const OBSERVATION_PAGE_SIZE = 20;
 const SUMMARY_PAGE_SIZE = 50;
@@ -128,28 +73,8 @@ function renderIntoFeedMount(mount: HTMLElement, content: ComponentChildren) {
 	render(h(TooltipProvider, null, content), mount);
 }
 
-function feedScopeLabel(scope: string): string {
-	if (scope === "mine") return " · my memories";
-	if (scope === "theirs") return " · other people";
-	return "";
-}
-
 function ProvenanceChip({ label, variant = "" }: { label: string; variant?: string }) {
 	return h(Chip, { variant: "provenance", tone: variant || undefined }, label);
-}
-
-function trustStateLabel(trustState: string): string {
-	if (trustState === "legacy_unknown") return "legacy provenance";
-	if (trustState === "unreviewed") return "unreviewed";
-	return trustState.replace(/_/g, " ");
-}
-
-function authorLabel(item: FeedItem): string {
-	if (item?.owned_by_self === true) return "You";
-	const actorId = String(item.actor_id || "").trim();
-	const actorName = String(item.actor_display_name || "").trim();
-	if (actorId && actorId === state.lastStatsPayload?.identity?.actor_id) return "You";
-	return actorName || actorId || "Unknown author";
 }
 
 function resetPagination(project: string) {
@@ -189,25 +114,6 @@ function pageNextOffset(payload: api.PaginatedResponse, count: number): number {
 
 function hasMorePages(): boolean {
 	return observationHasMore || summaryHasMore;
-}
-
-function mergeFeedItems(currentItems: FeedItem[], incomingItems: FeedItem[]): FeedItem[] {
-	const byKey = new Map<string, FeedItem>();
-	currentItems.forEach((item) => {
-		byKey.set(itemKey(item), item);
-	});
-	incomingItems.forEach((item) => {
-		byKey.set(itemKey(item), item);
-	});
-	return Array.from(byKey.values()).sort((a, b) => {
-		return new Date(b.created_at || 0).getTime() - new Date(a.created_at || 0).getTime();
-	});
-}
-
-function mergeRefreshFeedItems(currentItems: FeedItem[], firstPageItems: FeedItem[]): FeedItem[] {
-	const firstPageKeys = new Set(firstPageItems.map(itemKey));
-	const olderItems = currentItems.filter((item) => !firstPageKeys.has(itemKey(item)));
-	return mergeFeedItems(olderItems, firstPageItems);
 }
 
 export function syncInspectorQueryDraft(options: {
